Prevent frozenstep and netherstep blocks from breaking
Currently there is unexpected behaviour which would occur when a frozenstep or netherstep spawned block would be destroyed by explosives (like tnt) or moved by pistons. This commit aims to change that.

Note: this code should be optimised, as it introduces lag at following rates:
 -Addional piston lag when retraction or extending a piston proprtional to the blocks spawned by the frozenstep and netherstep enchantments
-Addional lag on explosions proprtional to the blocks spawned by the frozenstep and netherstep enchantments multiplied by the blocks cleared in the explosion

This commit also bumps the version of the plugin in the plugin.yml file to 1.6.3 as new methods have been added as well as to make the pom.xml and the plugin.yml versions the same