Remove builder pattern from Zenchantment
A new interface, Constructor, replaces the need for the builder pattern.
Calling 'defaults' for a Zenchantment required an instance of the class in the first place, which sort of defeated the point of the pattern.
Instead, the aim is to replace this system with a factory for creating Zenchantments with different properties.